corrected wallet transaction

+++ main/models.py
@@ -169,7 +169,7 @@ class Rental(models.Model):
     
     start_date = models.DateField()
     end_date = models.DateField()
-    total_price = models.IntegerField()
+    total_price = models.DecimalField(max_digits=10, decimal_places=2)
     
     full_name = models.CharField(max_length=150)
     age = models.IntegerField()
